+++ src/app/account/settings/tabs/account-tab/account-tab.component.ts
@@ -5,18 +5,24 @@
  * included in the LICENSE file.
  */
 
-import { ChangeDetectionStrategy, Component, inject } from "@angular/core";
+import { ChangeDetectionStrategy, Component, inject, Input, OnInit } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
+import { AbstractControl, FormBuilder, FormGroup, Validators } from "@angular/forms";
 import { ParamMap, Router } from "@angular/router";
 import { ToastrService } from "ngx-toastr";
 import { AccountService } from "src/app/account/account.service";
+import { AccountWithEmailFragment } from "src/app/api/kamu.graphql.interface";
 import { LoggedUserService } from "src/app/auth/logged-user.service";
 import { BaseComponent } from "src/app/common/components/base.component";
 import { ModalService } from "src/app/common/components/modal/modal.service";
 import { promiseWithCatch } from "src/app/common/helpers/app.helpers";
+import RoutingResolvers from "src/app/common/resolvers/routing-resolvers";
 import { MaybeNullOrUndefined, MaybeUndefined } from "src/app/interface/app.types";
 import ProjectLinks from "src/app/project-links";
 import { NavigationService } from "src/app/services/navigation.service";
+import { ChangeUsernameFormType } from "./account-tab.types";
+import { ChangeAccountUsernameResult } from "../../account-settings.constants";
+import { AccountTabs } from "src/app/account/account.constants";
 
 @Component({
     selector: "app-account-tab",
@@ -24,13 +30,29 @@ import { NavigationService } from "src/app/services/navigation.service";
     styleUrls: ["./account-tab.component.scss"],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class AccountTabComponent extends BaseComponent {
+export class AccountTabComponent extends BaseComponent implements OnInit {
+    @Input(RoutingResolvers.ACCOUNT_SETTINGS_ACCOUNT_KEY) public account: AccountWithEmailFragment;
+
+    public changeUsernameForm: FormGroup<ChangeUsernameFormType>;
+
     private modalService = inject(ModalService);
     private loggedUserService = inject(LoggedUserService);
     private router = inject(Router);
     private accountService = inject(AccountService);
     private toastrService = inject(ToastrService);
     private navigationService = inject(NavigationService);
+    private fb = inject(FormBuilder);
+
+    public ngOnInit(): void {
+        this.changeUsernameForm = this.fb.nonNullable.group({
+            username: [
+                this.isOwnerPage && !this.accountName
+                    ? this.loggedUserService.currentlyLoggedInUser.accountName
+                    : (this.accountName as string),
+                [Validators.required, Validators.pattern(/^([a-zA-Z0-9][a-zA-Z0-9-]*)+(\.[a-zA-Z0-9][a-zA-Z0-9-]*)*$/)],
+            ],
+        });
+    }
 
     public deleteAccount(): void {
         const message = "Do you want to delete a account";
@@ -63,6 +85,48 @@ export class AccountTabComponent extends BaseComponent {
                 },
             }),
         );
+    }
+
+    public changeUsername(): void {
+        const message = "Do you want to change username";
+        promiseWithCatch(
+            this.modalService.error({
+                title: "Change username",
+                message: message,
+                yesButtonText: "Ok",
+                noButtonText: "Cancel",
+                handler: (ok) => {
+                    if (ok) {
+                        const accountName =
+                            this.isOwnerPage && !this.accountName
+                                ? this.loggedUserService.currentlyLoggedInUser.accountName
+                                : (this.accountName as string);
+                        this.accountService
+                            .changeAccountUsername({
+                                accountName,
+                                newName: this.changeUsernameForm.controls.username.value,
+                            })
+                            .pipe(takeUntilDestroyed(this.destroyRef))
+                            .subscribe((result: ChangeAccountUsernameResult) => {
+                                if (result.changed) {
+                                    if (this.isOwnerPage && this.loggedUserService.currentlyLoggedInUser) {
+                                        this.loggedUserService.changeUser({
+                                            ...this.loggedUserService.currentlyLoggedInUser,
+                                            accountName: result.name,
+                                        });
+                                    } else {
+                                        this.navigationService.navigateToOwnerView(result.name, AccountTabs.SETTINGS);
+                                    }
+                                }
+                            });
+                    }
+                },
+            }),
+        );
+    }
+
+    public get usernameControl(): AbstractControl {
+        return this.changeUsernameForm.controls.username;
     }
 
     public get isOwnerPage(): boolean {