+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added webhook subscriptions feature
+- Support for account renaming
 
 ## [0.49.1] - 2025-06-05
 ### Added